financial-document-analyzer/requirements.txt
pdfplumber
google-genai
